requirements.txt
flask
flask-cors
pyyaml
pyjwt[crypto]
requests
sqlalchemy
psycopg2-binary
alembic
